Update RFC 0005 (Remove ManagedZone)
We will no longer support the ManagedZone API as part of MGC/Kuadrant.
This API was added at a time when we expected to create and manage many
hosted zones in the target DNS provider. This is no longer the case and
has no valid position in the future plans for a unified kuadrant.